+++ src/app/api/upload/route.js
@@ -11,8 +11,8 @@ export async function POST(req) {
         const s3Client = new S3Client({
             region: 'us-east-1',
             credentials: {
-                accessKeyId: process.env.AWS_ACCESS_KEY,
-                secretAccessKey: process.env.AWS_SECRET_KEY,
+                accessKeyId: process.env.AWS_ACCESSKEY,
+                secretAccessKey: process.env.AWS_SECRETKEY,
             }
 
         })
