+++ src/components/menu/MenuItem.js
@@ -1,3 +1,4 @@
+import Image from "next/image";
 
 
 
@@ -5,7 +6,7 @@
 const MenuItem = () => {
     return ( 
         <div className='bg-pink-100 p-8 rounded-lg text-center hover:bg-red-100 hover:shadow-xl hover:shadow-black/20 transition-all'>
-            <img src="/sprinkles.png" alt="donut" />
+            <Image src="/sprinkles.png" width={300} height={240} alt="donut" />
             <h4 className='text-primary font-semibold'>Sprinkles Donut</h4>
             <p className='mt-4'>$25 Per Dozen </p>
             <button className="mt-4 bg-primary rounded-full text-white px-6 py-2">Add to cart</button>
